pyproject.toml
[tool.poetry]
name = "sitg_ex_travel"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
annotated-types = "0.7.0"
anyio = "4.4.0"
certifi = "2024.7.4"
charset-normalizer = "3.4.0"
click = "8.1.7"
distro = "1.9.0"
dnspython = "2.6.1"
email-validator = "2.2.0"
fastapi = "0.111.1"
fastapi-cli = "0.0.4"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
Jinja2 = "3.1.4"
markdown-it-py = "3.0.0"
MarkupSafe = "2.1.5"
mdurl = "0.1.2"
openai = "1.37.1"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
Pygments = "2.18.0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
PyYAML = "6.0.1"
requests = "2.32.3"
rich = "13.7.1"
shellingham = "1.5.4"
sniffio = "1.3.1"
starlette = "0.37.2"
tqdm = "4.66.4"
typer = "0.12.3"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
uvicorn = "0.30.3"
uvloop = "0.19.0"
watchfiles = "0.22.0"
websockets = "12.0"
google = "3.0.0"
google-cloud-aiplatform = "^1.91.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
